+++ package.json
@@ -9,6 +9,8 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@expo/webpack-config": "^0.17.0",
+    "@reduxjs/toolkit": "^1.8.4",
     "expo": "~46.0.8",
     "expo-splash-screen": "~0.16.1",
     "expo-status-bar": "~1.4.0",
@@ -19,8 +21,7 @@
     "react-native": "0.69.4",
     "react-native-web": "~0.18.7",
     "react-redux": "^8.0.2",
-    "redux": "^4.2.0",
-    "@expo/webpack-config": "^0.17.0"
+    "redux": "^4.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9"
