+++ src/components/auth/authSlice.js
@@ -1,0 +1,19 @@
+import { createSlice } from '@reduxjs/toolkit';
+import { auth } from '../../firebase';
+
+const initialState = {
+  isLoggedIn: false,
+};
+
+export const authSlice = createSlice({
+  name: 'auth',
+  initialState,
+  reducers: {
+    setIsLoggedIn(state, action) {
+      state.isLoggedIn = action.payload;
+    },
+  },
+});
+
+export const { setIsLoggedIn } = authSlice.actions;
+export default authSlice.reducer;
